+++ src/balances/balances.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { BalancesService } from './balances.service';
 import { BalancesController } from './balances.controller';
-import { SafeTransactionModule } from '../services/safe-transaction/safe-transaction.module';
+import { TransactionServiceModule } from '../services/transaction-service/transaction-service.module';
 
 @Module({
   controllers: [BalancesController],
   providers: [BalancesService],
-  imports: [SafeTransactionModule],
+  imports: [TransactionServiceModule],
 })
 export class BalancesModule {}
