+++ src/services/transaction-service/transaction-service.service.spec.ts
@@ -1,7 +1,7 @@
 import { HttpService } from '@nestjs/axios';
 import { Balance } from './entities/balance.entity';
 import { HttpErrorHandler } from '../errors/http-error-handler';
-import { SafeTransactionService } from './safe-transaction.service';
+import { TransactionService } from './transaction-service.service';
 
 const BALANCES: Balance[] = [
   { tokenAddress: 'tokenAddress1', balance: 'balanceStr1' },
@@ -16,8 +16,8 @@ const mockHttpErrorHandler = {
   handle: jest.fn(),
 } as unknown as HttpErrorHandler;
 
-describe('SafeTransactionService', () => {
-  const service: SafeTransactionService = new SafeTransactionService(
+describe('TransactionService', () => {
+  const service: TransactionService = new TransactionService(
     'baseUrl',
     mockHttpService,
     mockHttpErrorHandler,
